Financial_Project/BackEnd/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-seed==0.3.1
djangorestframework==3.15.2
exceptiongroup==1.2.0
executing==2.0.1
Faker==30.4.0
fonttools==4.53.1
idna==3.10
importlib_resources==6.4.5
ipython==8.18.1
jedi==0.19.1
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.6
numpy==1.26.3
oauthlib==3.2.2
packaging==24.1
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.15.0
sqlparse==0.5.1
stack-data==0.6.3
toposort==1.10
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
scikit-learn==1.5.2
joblib==1.4.2
scipy==1.14.1
threadpoolctl==3.5.0